Rename C++ reserved keywords in headers for C++ compatibility
- export.h: rename 'delete' function pointer -> 'destroy', rename
  'export' parameter -> 'exp'; update all usages in block/export/,
  nbd/server.c
- char.h/char.c: rename 'private' parameter -> 'opaque'
- cxl_cdat.h: rename 'private' field -> 'priv'; update usages in
  hw/cxl/, hw/mem/, hw/pci-bridge/
- pci.h/pci.c: rename 'new' parameter -> 'cur'
- dump-arch.h: rename 'class' parameter -> 'elf_class'
- cxl_device.h: fix narrowing conversion (int -> uint64_t)
- block_int-io.h: fix narrowing conversion with explicit cast

+++ include/block/block_int-io.h
@@ -60,7 +60,7 @@ int coroutine_fn GRAPH_RDLOCK bdrv_co_pwritev_part(BdrvChild *child,
 static inline int coroutine_fn GRAPH_RDLOCK bdrv_co_pread(BdrvChild *child,
     int64_t offset, int64_t bytes, void *buf, BdrvRequestFlags flags)
 {
-    QEMUIOVector qiov = QEMU_IOVEC_INIT_BUF(qiov, buf, bytes);
+    QEMUIOVector qiov = QEMU_IOVEC_INIT_BUF(qiov, buf, (size_t)bytes);
     IO_CODE();
     assert_bdrv_graph_readable();
 
@@ -70,7 +70,7 @@ static inline int coroutine_fn GRAPH_RDLOCK bdrv_co_pread(BdrvChild *child,
 static inline int coroutine_fn GRAPH_RDLOCK bdrv_co_pwrite(BdrvChild *child,
     int64_t offset, int64_t bytes, const void *buf, BdrvRequestFlags flags)
 {
-    QEMUIOVector qiov = QEMU_IOVEC_INIT_BUF(qiov, buf, bytes);
+    QEMUIOVector qiov = QEMU_IOVEC_INIT_BUF(qiov, buf, (size_t)bytes);
     IO_CODE();
     assert_bdrv_graph_readable();
